+++ cuda_core/cuda/core/experimental/_context.py
@@ -5,7 +5,7 @@
 from dataclasses import dataclass
 
 from cuda.core.experimental._clear_error_support import assert_type
-from cuda.core.experimental._utils import driver
+from cuda.core.experimental._utils.cuda_utils import driver
 
 
 @dataclass
