+++ cuda_core/tests/test_stream.py
@@ -11,7 +11,7 @@ import pytest
 from cuda.core.experimental import Device, Stream, StreamOptions
 from cuda.core.experimental._event import Event
 from cuda.core.experimental._stream import LEGACY_DEFAULT_STREAM, PER_THREAD_DEFAULT_STREAM, default_stream
-from cuda.core.experimental._utils import driver
+from cuda.core.experimental._utils.cuda_utils import driver
 
 
 def test_stream_init_disabled():
